netcdf: fix NetCDFCRSTest to build with Java 13+
`NetCDFCRSTest`'s `setFinalStaticField()` was using reflection to
change the value of the statically initialized `NetCDFCRSUtilities.CONVERT_AXIS_KM`
boolean, but building with Java 13+ leads to a "no such field
'modifiers'" error.

The flag is final and picks the value of the
`org.geotools.coverage.io.netcdf.convertAxis.km` System property to be
returned by the static `NetCDFCRSUtilities.isConvertAxisKm()` utility
method.

Change the utility method to pick up the System property on the fly
instead, maintaining the default value of `false`;